+++ src/main/java/com/sion/concertbooking/infrastructure/config/JpaConfiguration.java
@@ -10,7 +10,7 @@ import org.springframework.transaction.PlatformTransactionManager;
 @Configuration
 @EnableJpaAuditing
 @EnableJpaRepositories(basePackages = {
-        "com.sion.concertbooking.infrastructure.repository",
+        "com.sion.concertbooking.infrastructure.impl",
         "com.sion.concertbooking.domain.repository"
 })
 public class JpaConfiguration {
