+++ src/main/java/com/sion/concertbooking/infrastructure/impl/ConcertRepositoryImpl.java
@@ -1,7 +1,8 @@
-package com.sion.concertbooking.infrastructure.repository;
+package com.sion.concertbooking.infrastructure.impl;
 
 import com.sion.concertbooking.domain.concert.Concert;
 import com.sion.concertbooking.domain.concert.ConcertRepository;
+import com.sion.concertbooking.infrastructure.jpa.ConcertJpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
