+++ src/main/java/com/sion/concertbooking/infrastructure/impl/ConcertScheduleRepositoryImpl.java
@@ -1,7 +1,8 @@
-package com.sion.concertbooking.infrastructure.repository;
+package com.sion.concertbooking.infrastructure.impl;
 
 import com.sion.concertbooking.domain.concertschedule.ConcertSchedule;
 import com.sion.concertbooking.domain.concertschedule.ConcertScheduleRepository;
+import com.sion.concertbooking.infrastructure.jpa.ConcertScheduleJpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
