+++ src/main/java/com/sion/concertbooking/infrastructure/impl/PointRepositoryImpl.java
@@ -1,7 +1,8 @@
-package com.sion.concertbooking.infrastructure.repository;
+package com.sion.concertbooking.infrastructure.impl;
 
 import com.sion.concertbooking.domain.point.Point;
 import com.sion.concertbooking.domain.point.PointRepository;
+import com.sion.concertbooking.infrastructure.jpa.PointJpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
