+++ src/main/java/com/sion/concertbooking/infrastructure/jpa/ConcertJpaRepository.java
@@ -1,4 +1,4 @@
-package com.sion.concertbooking.infrastructure.repository;
+package com.sion.concertbooking.infrastructure.jpa;
 
 import com.sion.concertbooking.domain.concert.Concert;
 import org.springframework.data.jpa.repository.JpaRepository;
